+++ src/main/java/com/programmers/jpa_board/post/application/PostServiceImpl.java
@@ -1,12 +1,12 @@
 package com.programmers.jpa_board.post.application;
 
-import com.programmers.jpa_board.global.converter.BoardConverter;
 import com.programmers.jpa_board.post.domain.Post;
 import com.programmers.jpa_board.post.domain.dto.request.CreatePostRequest;
 import com.programmers.jpa_board.post.domain.dto.request.UpdatePostRequest;
 import com.programmers.jpa_board.post.domain.dto.response.PostResponse;
 import com.programmers.jpa_board.post.exception.NotFoundPostException;
 import com.programmers.jpa_board.post.infra.PostRepository;
+import com.programmers.jpa_board.post.util.PostConverter;
 import com.programmers.jpa_board.user.application.UserProviderService;
 import com.programmers.jpa_board.user.domain.User;
 import org.springframework.data.domain.Page;
@@ -21,9 +21,9 @@ public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
     private final UserProviderService userService;
-    private final BoardConverter converter;
+    private final PostConverter converter;
 
-    public PostServiceImpl(PostRepository postRepository, UserProviderService userService, BoardConverter converter) {
+    public PostServiceImpl(PostRepository postRepository, UserProviderService userService, PostConverter converter) {
         this.postRepository = postRepository;
         this.userService = userService;
         this.converter = converter;
