+++ src/main/java/com/programmers/jpa_board/user/application/UserServiceImpl.java
@@ -1,11 +1,11 @@
 package com.programmers.jpa_board.user.application;
 
-import com.programmers.jpa_board.global.converter.BoardConverter;
 import com.programmers.jpa_board.user.domain.User;
 import com.programmers.jpa_board.user.domain.dto.request.CreateUserRequest;
 import com.programmers.jpa_board.user.domain.dto.response.UserResponse;
 import com.programmers.jpa_board.user.exception.NotFoundUserException;
 import com.programmers.jpa_board.user.infra.UserRepository;
+import com.programmers.jpa_board.user.util.UserConverter;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,9 +15,9 @@ public class UserServiceImpl implements UserProviderService {
     private static final String NOT_FOUND_USER = "회원이 존재하지 않습니다.";
 
     private final UserRepository userRepository;
-    private final BoardConverter converter;
+    private final UserConverter converter;
 
-    public UserServiceImpl(UserRepository userRepository, BoardConverter converter) {
+    public UserServiceImpl(UserRepository userRepository, UserConverter converter) {
         this.userRepository = userRepository;
         this.converter = converter;
     }
